requirements.txt
streamlit
PyPDF2
python-docx
spacy
langchain
openai
pinecone-client
python-dotenv
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz